+++ src/icalendar/__init__.py
@@ -9,7 +9,7 @@ from icalendar.prop import vBinary, vBoolean, vCalAddress, vDatetime, vDate, \
      vGeo, vUTCOffset, TypesFactory
 
 # useful tzinfo subclasses
-from icalendar.prop import FixedOffset, UTC, LocalTimezone
+from icalendar.prop import FixedOffset, LocalTimezone
 
 # Parameters and helper methods for splitting and joining string with escaped
 # chars.
@@ -22,6 +22,6 @@ __all__ = [
     vDDDTypes, vDuration, vFloat, vInt, vPeriod,
     vWeekday, vFrequency, vRecur, vText, vTime, vUri,
     vGeo, vUTCOffset, TypesFactory,
-    FixedOffset, UTC, LocalTimezone,
+    FixedOffset, LocalTimezone,
     Parameters, q_split, q_join,
 ]